requirements.txt
python-telegram-bot==20.6
supabase-auth==2.12.3
supabase-functions==0.10.1
requests
BackgroundScheduler
fastapi>=0.111
uvicorn[standard]>=0.30
python-dotenv>=1.0
supabase>=2.6.0
httpx>=0.28,<0.29

# 1) Clean out conflicting versions
#pip uninstall -y supabase supabase-auth supabase-functions httpx postgrest storage3 gotrue supafunc

# 2) Reinstall a consistent set (tested combo)
#pip install "httpx[http2]==0.28.1" "supabase==2.18.1" "supabase-auth==2.12.3" "supabase-functions==0.10.1"
